Clean up. *Need fix of onClick button
